added dodge charger after changes requested
