--- PyGamesScripts/Dodge-Charger/Music/Important.txt
@@ -1,4 +1,0 @@
-Cherry Metal by Arthur Vyncke | https://soundcloud.com/arthurvost
-Creative Commons Attribution-ShareAlike 3.0 Unported
-https://creativecommons.org/licenses/by-sa/3.0/deed.en_US
-Music promoted by https://www.chosic.com/
