+++ PyGamesScripts/Dodge-Charger/README.md
@@ -1,21 +1,39 @@
 # Dodge Charger Game
 - This game has made with pygame module of python. Here you have to drive your car on the highway and stay alive as long as you can by dodging your car.
+
 - #### Features
     - You can collect coins.
     - Use them to buy new vehicles.
     - Collect heart to increase vehicles's health.
-- #### Screenshots
- - ![Game-Image](Demo-Images\demo_img1.png)
- - ![Garage-Image](Demo-Images\demoj_img2.png)
 
-# Required
-- python -- [Download python](https://www.python.org/downloads/)
-- pygame --  ```pip install pygame```
-- all other modules have already come with python.
+- #### Screenshots
+ - ![Game-Image](Images\demo_img1.png)
+ - ![Garage-Image](Images\demoj_img2.png)
+
+# How to install the game locally?
+1. Install the python [Download python](https://www.python.org/downloads/)
+2. **Windows** Type/Paste this on command prompt ```pip install -r requirements.txt```
+3. All other modules that were used have already come with python.
+
+# Project Structure
+1. Assets - Contains all the images that have used in this game.
+2. Images - Contains screenshots of the game for README file
+3. Related_Files
+    1. A database(MainDB.db) file to save the coins data and how many cars have bought
+    and which car is currently equiped.
+    2. Python file(garage_menu.py) contains the code for the garage screen in the game.
+    3. Python file(main_game.py) contains the code of the main game.
+4. Sounds - Contains all the sounds like button click etc.
+5. dodge_charger.py contains the code of the main menu.
+6. README.md file - For explanation
+7. requirements.txt - module you have to download to run the game.
 
 # How to play?
-- Click on the start_menu file.
+- Click on the dodge_charger.py file
 - And Enjoy :rocket
 
 # Tip
 - If you don't collect heart then the speed of enemy's car won't increase. :+1
+
+# Author
+Mudit
